refactor(docs): input dx
