+++ apps/docs/content/components/input/clear-button.raw.jsx
@@ -1,0 +1,17 @@
+import {Input} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Input
+      isClearable
+      className="max-w-xs"
+      defaultValue="junior@nextui.org"
+      label="Email"
+      placeholder="Enter your email"
+      type="email"
+      variant="bordered"
+      // eslint-disable-next-line no-console
+      onClear={() => console.log("input cleared")}
+    />
+  );
+}
