+++ apps/docs/content/components/input/clear-button.ts
@@ -1,19 +1,4 @@
-const App = `import {Input} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Input
-      isClearable
-      type="email"
-      label="Email"
-      variant="bordered"
-      placeholder="Enter your email"
-      defaultValue="junior@nextui.org"
-      onClear={() => console.log("input cleared")}
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./clear-button.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
