+++ apps/docs/content/components/input/colors.ts
@@ -1,31 +1,4 @@
-const App = `import {Input} from "@nextui-org/react";
-
-export default function App() {
-  const colors = [
-    "default",
-    "primary",
-    "secondary",
-    "success",
-    "warning",
-    "danger",
-  ];
-
-  return (
-    <div className="w-full flex flex-row flex-wrap gap-4">
-      {colors.map((color) => (
-        <Input
-          key={color}
-          type="email"
-          color={color}
-          label="Email"
-          placeholder="Enter your email"
-          defaultValue="junior@nextui.org"
-          className="max-w-[220px]"
-        />
-      ))}
-    </div>
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
