+++ apps/docs/content/components/input/controlled.ts
@@ -1,20 +1,4 @@
-const App = `import {Input} from "@nextui-org/react";
-
-export default function App() {
-  const [value, setValue] = React.useState("");
-
-  return (
-    <div className="w-full flex flex-col gap-2 max-w-[240px]">
-      <Input
-        label="Email"
-        placeholder="Enter your email"
-        value={value}
-        onValueChange={setValue}
-      />
-      <p className="text-default-500 text-small">Input value: {value}</p>
-    </div>
-  );
-}`;
+import App from "./controlled.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
