+++ apps/docs/content/components/input/custom-impl.ts
@@ -1,164 +1,7 @@
-const SearchIcon = `export const SearchIcon = (props) => (
-  <svg
-    aria-hidden="true"
-    fill="none"
-    focusable="false"
-    height="1em"
-    role="presentation"
-    viewBox="0 0 24 24"
-    width="1em"
-    {...props}
-  >
-    <path
-      d="M11.5 21C16.7467 21 21 16.7467 21 11.5C21 6.25329 16.7467 2 11.5 2C6.25329 2 2 6.25329 2 11.5C2 16.7467 6.25329 21 11.5 21Z"
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="2"
-    />
-    <path
-      d="M22 22L20 20"
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeWidth="2"
-    />
-  </svg>
-);`;
-
-const CloseFilledIcon = `export const CloseFilledIcon = (props) => (
-  <svg
-    aria-hidden="true"
-    focusable="false"
-    height="1em"
-    role="presentation"
-    viewBox="0 0 24 24"
-    width="1em"
-    {...props}
-  >
-    <path
-      d="M12 2a10 10 0 1010 10A10.016 10.016 0 0012 2zm3.36 12.3a.754.754 0 010 1.06.748.748 0 01-1.06 0l-2.3-2.3-2.3 2.3a.748.748 0 01-1.06 0 .754.754 0 010-1.06l2.3-2.3-2.3-2.3A.75.75 0 019.7 8.64l2.3 2.3 2.3-2.3a.75.75 0 011.06 1.06l-2.3 2.3z"
-      fill="currentColor"
-    />
-  </svg>
-);`;
-
-const App = `import React, {forwardRef} from "react";
-import {useInput} from "@nextui-org/react";
-import {SearchIcon} from "./SearchIcon";
-import {CloseFilledIcon} from "./CloseFilledIcon";
-
-const styles = {
-  label: "text-black/50 dark:text-white/90",
-  input: [
-    "bg-transparent",
-    "text-black/90 dark:text-white/90",
-    "placeholder:text-default-700/50 dark:placeholder:text-white/60",
-  ],
-  innerWrapper: "bg-transparent",
-  inputWrapper: [
-    "shadow-xl",
-    "bg-default-200/50",
-    "dark:bg-default/60",
-    "backdrop-blur-xl",
-    "backdrop-saturate-200",
-    "hover:bg-default-200/70",
-    "focus-within:!bg-default-200/50",
-    "dark:hover:bg-default/70",
-    "dark:focus-within:!bg-default/60",
-    "!cursor-text",
-  ],
-};
-
-const MyInput = forwardRef((props, ref) => {
-  const {
-    Component,
-    label,
-    domRef,
-    description,
-    isClearable,
-    startContent,
-    endContent,
-    shouldLabelBeOutside,
-    shouldLabelBeInside,
-    errorMessage,
-    getBaseProps,
-    getLabelProps,
-    getInputProps,
-    getInnerWrapperProps,
-    getInputWrapperProps,
-    getDescriptionProps,
-    getErrorMessageProps,
-    getClearButtonProps,
-  } = useInput({
-    ...props,
-    ref,
-    // this is just for the example, the props bellow should be passed by the parent component
-    label: "Search",
-    type: "search",
-    placeholder: "Type to search...",
-    startContent: (
-      <SearchIcon className="text-black/50 mb-0.5 dark:text-white/90 text-slate-400 pointer-events-none flex-shrink-0" />
-    ),
-    // custom styles
-    classNames: {
-      ...styles,
-    },
-  });
-
-  const labelContent = <label {...getLabelProps()}>{label}</label>;
-
-  const end = React.useMemo(() => {
-    if (isClearable) {
-      return <span {...getClearButtonProps()}>{endContent || <CloseFilledIcon />}</span>;
-    }
-
-    return endContent;
-  }, [isClearable, getClearButtonProps]);
-
-  const innerWrapper = React.useMemo(() => {
-    if (startContent || end) {
-      return (
-        <div {...getInnerWrapperProps()}>
-          {startContent}
-          <input {...getInputProps()} />
-          {end}
-        </div>
-      );
-    }
-
-    return <input {...getInputProps()} />;
-  }, [startContent, end, getInputProps, getInnerWrapperProps]);
-
-  return (
-    <div className="w-[340px] h-[300px] px-8 rounded-2xl flex justify-center items-center bg-gradient-to-tr from-pink-500 to-yellow-500 text-white shadow-lg">
-      <Component {...getBaseProps()}>
-        {shouldLabelBeOutside ? labelContent : null}
-        <div
-          {...getInputWrapperProps()}
-          role="button"
-          onClick={() => {
-            domRef.current?.focus();
-          }}
-        >
-          {shouldLabelBeInside ? labelContent : null}
-          {innerWrapper}
-        </div>
-        {description && <div {...getDescriptionProps()}>{description}</div>}
-        {errorMessage && <div {...getErrorMessageProps()}>{errorMessage}</div>}
-      </Component>
-    </div>
-  );
-});
-
-MyInput.displayName = "MyInput";
-
-export default MyInput;`;
+import App from "./custom-impl.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
-  "/SearchIcon.jsx": SearchIcon,
-  "/CloseFilledIcon.jsx": CloseFilledIcon,
 };
 
 export default {
