+++ apps/docs/content/components/input/description.raw.jsx
@@ -1,0 +1,13 @@
+import {Input} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Input
+      className="max-w-xs"
+      defaultValue="junior@nextui.org"
+      description="We'll never share your email with anyone else."
+      label="Email"
+      type="email"
+    />
+  );
+}
