+++ apps/docs/content/components/input/description.ts
@@ -1,16 +1,4 @@
-const App = `import {Input} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Input
-      type="email"
-      label="Email"
-      defaultValue="junior@nextui.org"
-      description="We'll never share your email with anyone else."
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./description.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
