+++ apps/docs/content/components/input/error-message.ts
@@ -1,18 +1,4 @@
-const App = `import {Input} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Input
-      type="email"
-      label="Email"
-      variant="bordered"
-      defaultValue="junior2nextui.org"
-      isInvalid={true}
-      errorMessage="Please enter a valid email"
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./error-message.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
