+++ apps/docs/content/components/input/sizes.ts
@@ -1,20 +1,4 @@
-const App = `import {Input} from "@nextui-org/react";
-
-export default function App() {
-
-  const sizes = ["sm", "md", "lg"];
-
-  return (
-    <div className="w-full flex flex-col gap-4">
-      {sizes.map((size) => (
-        <div key={size} className="flex w-full flex-wrap md:flex-nowrap mb-6 md:mb-0 gap-4">
-          <Input size={size} type="email" label="Email" />
-          <Input size={size} type="email" label="Email" placeholder="Enter your email" />
-        </div>
-      ))}  
-    </div>  
-  );
-}`;
+import App from "./sizes.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
