+++ apps/docs/content/components/input/usage.ts
@@ -1,13 +1,4 @@
-const App = `import {Input} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-      <Input type="email" label="Email" />
-      <Input type="email" label="Email" placeholder="Enter your email" />
-    </div>
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
